+++ tosixinch/urlmap.py
@@ -130,7 +130,6 @@ class URL(object):
     """Unroot URLs."""
 
     MATCHER = re.compile('^https?://', flags=re.IGNORECASE)
-    INDEX = '_'
 
     def __init__(self, url):
         self._url = url
@@ -148,23 +147,13 @@ class URL(object):
             return True
         return False
 
-    def _add_index(self, url):
-        root, ext = posixpath.splitext(url)
-        if ext:
-            pass
-        elif '?' in url:
-            pass
-        else:
-            url = posixpath.join(url, self.INDEX)
-        return url
-
     def unroot(self):
         url = self.url
         url, _ = _split_fragment(url)
         url = self.MATCHER.sub('', url)
+        url = url.rstrip('/')
         if self._is_authority_only(url):
             return posixpath.join(url, 'index.html')
-        url = self._add_index(url)
         url = _url2path(url)
         return url
 
@@ -273,7 +262,7 @@ class Map(object):
     'map' here means to change somewhat neutral paths to actual filepaths.
     """
 
-    INDEX = URL.INDEX
+    INDEX = '_'
 
     def __init__(self, input_name, input_type=None):
         self._input_name = input_name
@@ -299,12 +288,25 @@ class Map(object):
             return FileURL(input_name)
         return Path(input_name)
 
+    def _add_index(self, url):
+        root, ext = posixpath.splitext(url)
+        if ext:
+            pass
+        elif '?' in url:
+            pass
+        else:
+            url = posixpath.join(url, self.INDEX)
+        return url
+
     def _map_name(self, name):
         # 'name' is always derived from a url, so it is ascii.
         for segment in name.split(self.sep):
             if len(segment) > 255:
                 self._hashed = True
                 return hashlib.sha1(name.encode('utf-8')).hexdigest()
+
+        if self.is_remote:
+            return self._add_index(name)
         return name
 
     @property